requirements.txt
deprecation==2.1.0
fpdf==1.7.2
img2pdf==0.4.4
lxml==4.9.1
packaging==21.3
pdf2image==1.16.0
pikepdf==6.2.4
Pillow==9.3.0
pip==22.0.4
pyparsing==3.0.9
setuptools==58.1.0
zipfile36